Some changes to support having the ZONEMD type defined in ldns.
Adding the ZONEMD type to the ldns library is relatively simple.
These changes allow ldns-zone-digest to work in that case.  It was
convenient to split zonemd_rr_pack() into _pack() and _create().
Similarly, it was convenient to split _find() into _find() and
_unpack().  Then zonemd_rr_update_digest() can call _unpack() and
_pack() instead of manipulating the RDFs itself.

The software now probes the ldns library runtime to see if it knows
about the ZONEMD record, and then behaves accordingly.